Restore ChoiceMap to inherit from Sample, and eliminate all pyright errors.
