+++ src/genjax/_src/generative_functions/combinators/dimap.py
@@ -204,9 +204,10 @@ class DimapCombinator(Generic[ArgTuple, R, S], GenerativeFunction[S]):
 
     def assess(
         self,
-        sample: ChoiceMap,
+        sample: Sample,
         args: tuple[Any, ...],
     ) -> tuple[Score, S]:
+        assert isinstance(sample, ChoiceMap)
         inner_args = self.argument_mapping(*args)
         w, inner_retval = self.inner.assess(sample, inner_args)
         retval = self.retval_mapping(inner_args, inner_retval)
